sla_manager/requirements.txt
Flask
requests
mysql-connector-python
prometheus-api-client